Refactor test folders for ed25519 (#364)
