+++ frost-ed25519/src/lib.rs
@@ -14,6 +14,8 @@ use sha2::{Digest, Sha512};
 use frost_core::frost;
 
 #[cfg(test)]
+mod helper_functions;
+#[cfg(test)]
 mod tests;
 
 // Re-exports in our public API
