+++ frost-ed25519/src/tests/batch_tests.rs
@@ -1,0 +1,17 @@
+use rand::thread_rng;
+
+use crate::*;
+
+#[test]
+fn check_batch_verify() {
+    let rng = thread_rng();
+
+    frost_core::tests::batch::batch_verify::<Ed25519Sha512, _>(rng);
+}
+
+#[test]
+fn check_bad_batch_verify() {
+    let rng = thread_rng();
+
+    frost_core::tests::batch::bad_batch_verify::<Ed25519Sha512, _>(rng);
+}
